# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AnalyzeResponse": {
        "additionalProperties": false,
        "properties": {
          "analyzed_requirement": {
            "$ref": "#/components/schemas/AnalyzedRequirement"
          },
          "change_summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChangeSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "clear_fields": {
            "items": {
              "$ref": "#/components/schemas/ClearFieldSuggestion"
            },
            "title": "Clear Fields",
            "type": "array"
          },
          "keywords": {
            "items": {
              "$ref": "#/components/schemas/KeywordSuggestion"
            },
            "title": "Keywords",
            "type": "array"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "requirement_id": {
            "format": "uuid",
            "title": "Requirement Id",
            "type": "string"
          },
          "reused_mappings": {
            "items": {
              "$ref": "#/components/schemas/ReusedMapping"
            },
            "title": "Reused Mappings",
            "type": "array"
          },
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "trace": {
            "$ref": "#/components/schemas/ProcessingTrace"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "requirement_id",
          "revision_id",
          "analyzed_requirement",
          "trace"
        ],
        "title": "AnalyzeResponse",
        "type": "object"
      },
      "AnalyzedRequirement": {
        "additionalProperties": false,
        "properties": {
          "business_goal": {
            "minLength": 1,
            "title": "Business Goal",
            "type": "string"
          },
          "constraints": {
            "items": {
              "type": "string"
            },
            "title": "Constraints",
            "type": "array"
          },
          "domain_context": {
            "items": {
              "type": "string"
            },
            "title": "Domain Context",
            "type": "array"
          },
          "negative_constraints": {
            "items": {
              "type": "string"
            },
            "title": "Negative Constraints",
            "type": "array"
          },
          "summary": {
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "target_output": {
            "default": "GDA",
            "title": "Target Output",
            "type": "string"
          }
        },
        "required": [
          "summary",
          "business_goal"
        ],
        "title": "AnalyzedRequirement",
        "type": "object"
      },
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
      "AssetReference": {
        "additionalProperties": false,
        "properties": {
          "asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "asset_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Name"
          },
          "source_system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source System"
          }
        },
        "title": "AssetReference",
        "type": "object"
      },
      "AttributeReference": {
        "additionalProperties": false,
        "properties": {
          "attribute_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribute Id"
          },
          "attribute_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribute Name"
          }
        },
        "title": "AttributeReference",
        "type": "object"
      },
...
      "ChangeSummary": {
        "additionalProperties": false,
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/RequirementChange"
            },
            "title": "Changes",
            "type": "array"
          }
        },
        "title": "ChangeSummary",
        "type": "object"
      },
...
      "DemoKnowledgeScenario": {
        "enum": [
          "empty",
          "complete_mapping",
          "partial_mapping",
          "no_evidence",
          "timeout",
          "invalid"
        ],
        "title": "DemoKnowledgeScenario",
        "type": "string"
      },
...
      "FeedbackDecision": {
        "enum": [
          "accept",
          "reject"
        ],
        "title": "FeedbackDecision",
        "type": "string"
      },
      "FeedbackTargetType": {
        "enum": [
          "asset",
          "attribute",
          "field_mapping"
        ],
        "title": "FeedbackTargetType",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InitialJobRequest": {
        "additionalProperties": false,
        "properties": {
          "business_domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Domain"
          },
          "knowledge_base_scenario": {
            "$ref": "#/components/schemas/DemoKnowledgeScenario",
            "default": "empty"
          },
          "requirement": {
            "minLength": 1,
            "title": "Requirement",
            "type": "string"
          },
          "target_gda": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Gda"
          }
        },
        "required": [
          "requirement"
        ],
        "title": "InitialJobRequest",
        "type": "object"
      },
      "JobResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "request_payload": {
            "additionalProperties": true,
            "title": "Request Payload",
            "type": "object"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalyzeResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "request_id",
          "status",
          "created_at",
          "updated_at",
          "request_payload"
        ],
        "title": "JobResponse",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "completed",
          "failed"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobSubmission": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "job_id",
          "request_id",
          "status"
        ],
        "title": "JobSubmission",
        "type": "object"
      },
...
      "ProcessingTrace": {
        "additionalProperties": false,
        "properties": {
          "knowledge_base_queries": {
            "items": {
              "type": "string"
            },
            "title": "Knowledge Base Queries",
            "type": "array"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "processing_time_ms": {
            "minimum": 0.0,
            "title": "Processing Time Ms",
            "type": "integer"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          }
        },
        "required": [
          "prompt_version",
          "model",
          "processing_time_ms"
        ],
        "title": "ProcessingTrace",
        "type": "object"
      },
...
      "RevisionResponse": {
        "additionalProperties": false,
        "properties": {
          "analyzed_requirement": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyzed Requirement"
          },
          "changes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Changes",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "feedback": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Feedback",
            "type": "array"
          },
          "full_requirement": {
            "title": "Full Requirement",
            "type": "string"
          },
          "output_snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Snapshot"
          },
          "requirement_id": {
            "format": "uuid",
            "title": "Requirement Id",
            "type": "string"
          },
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          },
          "revision_number": {
            "title": "Revision Number",
            "type": "integer"
          },
          "supplemental_information": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplemental Information"
          }
        },
        "required": [
          "requirement_id",
          "revision_id",
          "revision_number",
          "full_requirement",
          "supplemental_information",
          "analyzed_requirement",
          "changes",
          "feedback",
          "output_snapshot",
          "created_at"
        ],
        "title": "RevisionResponse",
        "type": "object"
      },
      "SearchFeedback": {
        "additionalProperties": false,
        "properties": {
          "asset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "attribute": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AttributeReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "candidate_id": {
            "minLength": 1,
            "title": "Candidate Id",
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/FeedbackDecision"
          },
          "field_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Name"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "target_type": {
            "$ref": "#/components/schemas/FeedbackTargetType"
          }
        },
        "required": [
          "candidate_id",
          "target_type",
          "decision"
        ],
        "title": "SearchFeedback",
        "type": "object"
      },
...
      "StageEventResponse": {
        "additionalProperties": false,
        "properties": {
          "duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Metadata",
            "type": "object"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "stage": {
            "$ref": "#/components/schemas/WorkflowStage"
          },
          "status": {
            "$ref": "#/components/schemas/StageStatus"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "request_id",
          "stage",
          "status",
          "sequence",
          "timestamp"
        ],
        "title": "StageEventResponse",
        "type": "object"
      },
      "StageStatus": {
        "enum": [
          "started",
          "completed",
          "failed"
        ],
        "title": "StageStatus",
        "type": "string"
      },
...
      "UpdateJobRequest": {
        "additionalProperties": false,
        "properties": {
          "knowledge_base_scenario": {
            "$ref": "#/components/schemas/DemoKnowledgeScenario",
            "default": "empty"
          },
          "requirement_id": {
            "format": "uuid",
            "title": "Requirement Id",
            "type": "string"
          },
          "search_feedback": {
            "items": {
              "$ref": "#/components/schemas/SearchFeedback"
            },
            "title": "Search Feedback",
            "type": "array"
          },
          "supplemental_information": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplemental Information"
          }
        },
        "required": [
          "requirement_id"
        ],
        "title": "UpdateJobRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkflowStage": {
        "enum": [
          "validating_input",
          "loading_revision",
          "updating_requirement",
          "analyzing_requirement",
          "searching_knowledge_base",
          "reconstructing_mappings",
          "calculating_confidence",
          "persisting_revision",
          "completed",
          "failed"
        ],
        "title": "WorkflowStage",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "Analyze Agent API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs/initial": {
      "post": {
        "operationId": "submit_initial_api_v1_jobs_initial_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitialJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmission"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Initial"
      }
    },
    "/api/v1/jobs/update": {
      "post": {
        "operationId": "submit_update_api_v1_jobs_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmission"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Update"
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job"
      }
    },
    "/api/v1/jobs/{job_id}/events": {
      "get": {
        "operationId": "stream_events_api_v1_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StageEventResponse"
                }
              },
              "text/event-stream": {}
            },
            "description": "Server-Sent Events containing stage and job payloads."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Events"
      }
    },
...
    "/api/v1/requirements/{requirement_id}/latest": {
      "get": {
        "operationId": "latest_revision_api_v1_requirements__requirement_id__latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "requirement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Requirement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest Revision"
      }
    },
    "/api/v1/requirements/{requirement_id}/revisions": {
      "get": {
        "operationId": "list_revisions_api_v1_requirements__requirement_id__revisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "requirement_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Requirement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RevisionResponse"
                  },
                  "title": "Response List Revisions Api V1 Requirements  Requirement Id  Revisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Revisions"
      }
    },
...
  }
}